Move exponential model into examples/standalone
