+++ examples/standalone/exponential_decay.py
